pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Sameer Ahmad <samimughal7786@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
pandas = "1.5.2"
jupyter = "1.0.0"
pyarrow = "10.0.1"
tqdm = "4.64.1"
plotly = "5.11.0"
requests = "2.28.1"
scikit-learn = "1.3.1"
xgboost = "2.0.0"
optuna = "3.4.0"
python-dotenv = "1.0.0"
lightgbm = "4.1.0"
twofish = "^0.3.0"
hopsworks = "3.4.4"
config = "^0.5.1"
setuptools = "^69.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
